LayoutTests/media/video-currentTime-expected.txt

TEST(video.currentTime == 0) OK
EVENT(load) TEST(video.currentTime == 0) OK
EVENT(play)
TEST(video.currentTime > 0) OK
END OF TEST
